backend/requirements.txt
fastapi
uvicorn
easyocr
pillow
python-multipart
numpy
opencv-python
